// AMM/SushiSwapV3/types.ts
export enum Fees {
    VERY_LOW = 100, // 0.01%
    LOW = 500,      // 0.05%
    MEDIUM = 3000,  // 0.3%
    BIG = 10000     // 1%
}
// ...
export type Chains = 'arbitrum' | 'polygon' | 'optimism' | 'ethereum' | 'avalanche' | 'bsc' | 'polygonTestnet' | 'arbitrumTestnet' |
                     'avalancheTestnet' | 'ethereumTestnet'